backend/requirements.txt
fastapi
uvicorn
torch
accelerate
transformers
diffusers
safetensors
imageio
imageio[ffmpeg]
opencv-python
xformers
huggingface_hub
Pillow
